requirements.txt
python-frontmatter>=1.0.0
pathlib>=1.0.1
pyyaml>=6.0.1
aiohttp>=3.8.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.3.0
jinja2>=3.0.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
regex>=2023.0.0
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0 